Refactor versioned apis
