+++ indigo_content_api/v1/serializers.py
